+++ R/kpi_groups.R
@@ -46,12 +46,12 @@ kpi_grp_extract_ids <- function(kpi_grp_df) {
 #' @export
 kpi_grp_unnest <- function(kpi_grp_df) {
   kpi_grp_df %>%
-    tidyr::unnest(cols = c(.data$members)) %>%
+    tidyr::unnest(cols = c("members")) %>%
     dplyr::select(
-      group_id = .data$id, id = .data$member_id,
-      group_title = .data$title, title = .data$member_title
+      group_id = "id", id = "member_id",
+      group_title = "title", title = "member_title"
     ) %>%
-    dplyr::select(.data$id, .data$title, .data$group_id, .data$group_title)
+    dplyr::select("id", "title", "group_id", "group_title")
 }